Split integration suite into test-systemd.py and test.py
verify.py was doing two jobs: installing redirect on a fresh systemd
host (test_start), and asserting API behaviour against the running
service. The pipeline now has two distinct phases (systemd install,
then docker deploy), so put each phase in its own file:

- test-systemd.py owns the tarball install plus a smoke index check
  and the apache/journal log-collection teardown.
- test.py (renamed from verify.py) holds the API suite that runs
  against the docker-deployed redirect.

Move the add_host_alias setup into an autouse session-scoped fixture
in conftest.py so both files get it without any /etc/hosts bash hack
in the Drone yaml. Drop the now-unused imports left in test.py and
the corresponding --deselect flags from the Drone commands.
